% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{indonesia_latlong_df}
\alias{indonesia_latlong_df}
\title{Coordinate of Several Regions in Indonesia}
\format{
A data frame with 4 observations and 3 variables:
\describe{
  \item{City}{Name of the region in Indonesia (character)}
  \item{latitude}{Latitude coordinate of the region (numeric)}
  \item{longitude}{Longitude coordinate of the region (numeric)}
}
}
\source{
Data taken from the gstar package version 0.1.0
}
\usage{
data(indonesia_latlong_df)
}
\description{
This dataset, indonesia_latlong_df, is a data frame containing the geographical coordinates
of several regions in Indonesia, specifically Semarang, Surakarta, Tegal, and Purwokerto.
It includes the city names and their respective latitude and longitude values.
}
\details{
The dataset name has been kept as 'indonesia_latlong_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndonesiAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
