% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indonesiapis-package.R
\docType{package}
\name{IndonesiAPIs}
\alias{IndonesiAPIs}
\alias{IndonesiAPIs-package}
\title{IndonesiAPIs: Access Indonesian Data via Public APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'Nager.Date', 'World Bank API', and 'REST Countries API', retrieving real-time or historical data related to Indonesia, such as holidays, economic indicators, and international demographic and geopolitical indicators. The package also includes a curated collection of open datasets focused on Indoesia, covering topics such as consumer prices, poverty probability, food prices by region, tourism destinations, and minimum wage statistics.
}
}
\details{
IndonesiAPIs: Access Indonesian Data via Public APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Indonesian Data via Public APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/indonesiapis}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
