% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_datasets_IndonesiAPIs.R
\name{view_datasets_IndonesiAPIs}
\alias{view_datasets_IndonesiAPIs}
\title{View Available Datasets in IndonesiAPIs}
\usage{
view_datasets_IndonesiAPIs()
}
\value{
A character vector with the names of the available datasets.
        If no datasets are found, it returns an empty character vector.
}
\description{
This function lists all datasets available in the 'IndonesiAPIs' package.
If the 'IndonesiAPIs' package is not loaded, it stops and shows an error message.
If no datasets are available, it returns a message and an empty vector.
}
\examples{
if (requireNamespace("IndonesiAPIs", quietly = TRUE)) {
  library(IndonesiAPIs)
  view_datasets_IndonesiAPIs()
}
}
